Create migration for Product, ProductImage, Ctagory tables

+++ OnlineStore-Api/Models/Category.cs
@@ -2,7 +2,7 @@
 
 public class Category
 {
-    public int CategoryID { get; set; }
+    public byte CategoryID { get; set; }
     public string Title { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
 }
